+++ pkg/providers/property/testdata/TestDataPropertyAccountHostnames/valid_with_hostname_filter.tf
@@ -1,0 +1,6 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+data "akamai_property_account_hostnames" "test" {
+  hostname = "example.com"
+}